Move util functions from logging submodule to optimagic.utilities, remove obsolete submodule
